M68K backend now correctly prints out a nop!

+++ arch/M68K/M68Kdasm.h
@@ -346,7 +346,7 @@ unsigned int m68k_is_valid_instruction(unsigned int instruction, unsigned int cp
 /* Disassemble 1 instruction using the epecified CPU type at pc.  Stores
  * disassembly in str_buff and returns the size of the instruction in bytes.
  */
-unsigned int m68k_disassemble(MCInst* inst, uint64_t pc, unsigned int cpu_type);
+unsigned int m68k_disassemble(MCInst* inst, unsigned int pc, unsigned int cpu_type);
 
 /* ======================================================================== */
 /* ============================== MAME STUFF ============================== */
